+++ src/test/java/org/sonar/plugins/delphi/pmd/profile/DefaultDelphiProfileTest.java
@@ -45,7 +45,7 @@ public class DefaultDelphiProfileTest {
         context.profile(DelphiLanguage.KEY, DefaultDelphiProfile.DEFAULT_PROFILE_NAME);
 
     List<BuiltInActiveRule> activeRules = profile.rules();
-    assertThat(activeRules).hasSize(51);
+    assertThat(activeRules).hasSize(52);
     assertThat(profile.name()).isEqualTo(DefaultDelphiProfile.DEFAULT_PROFILE_NAME);
 
     // Check that we use severity from the read rule and not default one.